+++ lib/src/result_monad_base.dart
@@ -82,10 +82,10 @@ class Result<T, E> {
   /// ```dart
   /// return doSomething()
   ///   .andThen((r1) => r1.doSomething1())
-  ///   .andThen((r2) => r2.doSomething2())
+  ///   .transform((r2) => r2.toString())
   ///   .andThen((r3) => r3.doSomething3())
   /// ```
-  Result<T2, dynamic> andThenSuccess<T2, E2>(T2 Function(T) thenFunction) {
+  Result<T2, dynamic> transform<T2, E2>(T2 Function(T) thenFunction) {
     if (isSuccess) {
       try {
         return Result.ok(thenFunction(_value));
@@ -95,6 +95,25 @@ class Result<T, E> {
     }
 
     return Result.error(_error);
+  }
+
+  /// <b>NOTE: This is old syntax preserved for backward compatibility. Instead
+  /// use the [transform] method</b>
+  ///
+  /// Executes the anonymous function passing the current value to it to support
+  /// operation chaining. The function returns a non-Result value therefore
+  /// this assumes that the operation is always a success. This is helpful for
+  /// chaining functions that don't have result monads.
+  ///
+  /// Example:
+  /// ```dart
+  /// return doSomething()
+  ///   .andThen((r1) => r1.doSomething1())
+  ///   .andThenSuccess((r2) => r2.doSomething2())
+  ///   .andThen((r3) => r3.doSomething3())
+  /// ```
+  Result<T2, dynamic> andThenSuccess<T2, E2>(T2 Function(T) thenFunction) {
+    return transform(thenFunction);
   }
 
   /// Executes the anonymous async function passed in on the current value. Due
@@ -134,7 +153,40 @@ class Result<T, E> {
   /// final asyncResult = await doSomething()
   ///   .andThen((r1) => r1.doSomething1())
   ///   .andThen((r2) => r2.doSomething2())
-  ///   .andThenAsync((r3) async  => r3.doSomething3())
+  ///   .transform((r3) async  => r3.toString())
+  ///
+  /// return asyncResult
+  ///   .andThen((r4) => r4.doSomething4())
+  ///   .andThen((r5) => r5.doSomething5())
+  /// ```
+  FutureResult<T2, dynamic> transformAsync<T2, E2>(
+      Future<T2> Function(T) thenFunction) async {
+    if (isSuccess) {
+      try {
+        final result = await thenFunction(_value);
+        return Result.ok(result);
+      } catch (e) {
+        return Result.error(e);
+      }
+    }
+
+    return Result.error(_error);
+  }
+
+  /// <b>NOTE: This is old syntax preserved for backward compatibility. Instead
+  /// use the [transformAsync] method</b>
+  ///
+  /// Executes the anonymous async function passed in on the current value. Due
+  /// to Dart syntax peculiarities it may be cleaner to work with interim
+  /// results.The function returns a non-Result value therefore
+  //  this assumes that the operation is always a success. This is helpful for
+  //  chaining functions that don't have result monads.
+  ///
+  /// ```dart
+  /// final asyncResult = await doSomething()
+  ///   .andThen((r1) => r1.doSomething1())
+  ///   .andThen((r2) => r2.doSomething2())
+  ///   .andThenSuccessAsync((r3) async  => r3.toString())
   ///
   /// return asyncResult
   ///   .andThen((r4) => r4.doSomething4())
@@ -308,7 +360,8 @@ class Result<T, E> {
   ///   .withAsync((r2) => print(r2))
   ///   .andThen((r2) => r2.doSomething3())
   /// ```
-  FutureResult<T, dynamic> withResultAsync(Future<void> Function(T) withFunction) async {
+  FutureResult<T, dynamic> withResultAsync(
+      Future<void> Function(T) withFunction) async {
     if (isSuccess) {
       try {
         await withFunction(_value);
@@ -320,7 +373,6 @@ class Result<T, E> {
 
     return Result.error(_error);
   }
-
 
   @override
   String toString() {